+++ vfxGraph/vfxNodes/vfxNodeSpectrum1D.cpp
@@ -183,7 +183,7 @@ void VfxNodeSpectrum1D::tick(const float dt)
 				{
 					const float r = rreal[x];
 					const float i = rimag[x];
-					const float s = std::sqrtf(r * r + i * i);
+					const float s = std::sqrt(r * r + i * i);
 					
 					rreal[x] = s * scale;
 				}
